Rotate drivers locally
